pyproject.toml
[project]
name = "neofish"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "anthropic>=0.84.0",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.135.1",
    "numpy>=2.3.4",
    "openai>=2.26.0",
    "playwright>=1.58.0",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
    "websockets>=16.0",
]
